Move constants in kostal_plenticore
